--- .npmignore
@@ -1,3 +1,0 @@
-src
-docs
-webpack.config.cjs
