+++ babel.config.js
@@ -1,14 +1,13 @@
-var commonPresets = ['@babel/preset-react', '@babel/preset-typescript'];
+const commonPresets = ['@babel/preset-typescript']
 
 module.exports = (api) => {
   return api.env('test') ? {
     presets: [
       '@babel/preset-env',
-    ].concat(commonPresets),
-  } : {
-    ignore: [
-      '**/*.test.js',
+      '@babel/preset-react',
+      ...commonPresets,
     ],
+  } : {
     plugins: [
       ...(process.env.BUILD_TYPE === 'cjs' ? ['@babel/plugin-transform-modules-commonjs'] : []),
       'babel-plugin-jsx-remove-data-test-id',
