+++ jest.config.js
@@ -1,4 +1,4 @@
 module.exports = {
   rootDir: 'tests',
-  setupFilesAfterEnv: ['./jest.setup.js'],
+  setupFilesAfterEnv: ['./jest.setup.ts'],
 }
