+++ package.json
@@ -13,7 +13,8 @@
   },
   "files": [
     "dist-cjs",
-    "dist-esm"
+    "dist-esm",
+    "dist-types"
   ],
   "keywords": [
     "animated",
@@ -30,19 +31,19 @@
   ],
   "main": "dist-cjs",
   "module": "dist-esm",
-  "types": "./dist-types/index.d.ts",
+  "types": "dist-types/index.d.ts",
   "sideEffects": false,
   "scripts": {
-    "type-check": "tsc --noEmit",
-    "type-check:watch": "npm run type-check -- --watch",
+    "build:cjs": "BUILD_TYPE=cjs babel src --extensions '.ts,.tsx' --out-dir dist-cjs",
+    "build:esm": "babel src --extensions '.ts,.tsx' --out-dir dist-esm",
+    "build:types": "tsc --emitDeclarationOnly",
+    "build": "run-p build:cjs build:esm build:types",
     "prebuild": "rimraf dist-*",
-    "build": "run-p build:types build:cjs build:esm",
-    "build:cjs": "BUILD_TYPE=cjs webpack",
-    "build:esm": "webpack",
-    "build:types": "tsc --noEmit false --emitDeclarationOnly --outDir ./dist-types",
     "prepublishOnly": "npm test && npm run build",
     "test:watch": "jest --watch",
-    "test": "jest"
+    "test": "jest",
+    "type-check:watch": "npm run type-check -- --watch",
+    "type-check": "tsc --noEmit"
   },
   "dependencies": {},
   "devDependencies": {
@@ -54,18 +55,13 @@
     "@babel/preset-typescript": "^7.9.0",
     "@testing-library/jest-dom": "^5.3.0",
     "@testing-library/react": "^10.0.1",
-    "@types/react": "^16.9.26",
-    "@types/react-dom": "^16.9.5",
-    "babel-loader": "^8.1.0",
     "babel-plugin-jsx-remove-data-test-id": "^2.1.3",
     "jest": "^25.2.3",
     "npm-run-all": "^4.1.5",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "rimraf": "^3.0.2",
-    "typescript": "^3.8.3",
-    "webpack": "^4.42.1",
-    "webpack-cli": "^3.3.11"
+    "typescript": "^3.8.3"
   },
   "peerDependencies": {
     "react": "^16.8.0"
