+++ src/Fade.tsx
@@ -1,6 +1,6 @@
 import { Burger } from './Burger'
 import React, { FunctionComponent } from 'react'
-import { CommonBurgerProps } from './common-types'
+import { CommonBurgerProps } from './'
 
 export const Fade = ((props) => (
   <Burger {...props} render={(o) => (
