--- tests/jest.setup.cjs
@@ -1,1 +1,0 @@
-require('@testing-library/jest-dom');
