+++ tsconfig.json
@@ -1,31 +1,16 @@
 {
   "compilerOptions": {
-    "module": "esnext",
-    "target": "es6",
-    "jsx": "react",
     "declaration": true,
-    "rootDir": "./src",
+    "esModuleInterop": true,
+    "jsx": "react",
+    "module": "ESNext",
+    "moduleResolution": "Node",
+    "outDir": "./dist-types",
     "removeComments": false,
     "strict": true,
-    "noImplicitAny": true,
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "moduleResolution": "node",
-    "noEmit": true,
-    "lib": [
-      "es6",
-      "dom",
-      "es2016",
-      "es2017"
-    ]
+    "target": "ESNext",
   },
-  "exclude": [
-    "node_modules",
-    "tests",
-    "docs"
-  ],
   "include": [
-    "src/*.ts",
-    "src/*.tsx"
+    "src"
   ]
 }
